refactor: change MAX_FF_EFFECTS type from usize to u32 for consistency

+++ src/evdev_helpers.rs
@@ -6,7 +6,7 @@ use gilrs::{Axis, Button};
 use std::error::Error;
 
 /// Shared constant for the maximum number of force feedback effects.
-pub const MAX_FF_EFFECTS: usize = 16;
+pub const MAX_FF_EFFECTS: u32 = 16;
 
 // --- Scaling Helper Functions ---
 
@@ -105,7 +105,7 @@ pub fn create_virtual_gamepad(info: &VirtualGamepadInfo) -> Result<VirtualDevice
         // Add more effect codes if needed
     ]);
     builder = builder.with_ff(&ff_effects)?;
-    builder = builder.with_ff_effects_max(MAX_FF_EFFECTS as u32);
+    builder = builder.with_ff_effects_max(MAX_FF_EFFECTS);
 
     Ok(builder.build()?)
 }
